requirements.txt
absl-py==2.2.0
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
folium==0.19.5
fonttools==4.56.0
fpdf==1.7.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
holidays==0.69
huepy==1.2.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipwhois==1.3.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonpickle==4.0.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.0
kiwisolver==1.4.8
kombu==5.5.0
libclang==18.1.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
narwhals==1.31.0
networkx==3.4.2
numpy==2.1.3
opt_einsum==3.4.0
optree==0.14.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
plotly==6.0.1
prompt_toolkit==3.0.50
prophet==1.1.6
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyee==11.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.2
pyppeteer==2.0.0
pyshark==0.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-whois==0.9.5
pytz==2025.1
pyvis==0.3.2
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.23.1
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.30.0
setuptools==77.0.3
shodan==1.31.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
stanio==0.5.1
starlette==0.46.1
streamlit==1.43.2
streamlit_folium==0.24.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
threadpoolctl==3.6.0
tldextract==5.1.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
uvicorn==0.34.0
vine==5.1.0
virustotal-python==1.1.0
wappalyzer==1.0.11
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
XlsxWriter==3.2.2
xyzservices==2025.1.0
zipp==3.21.0
